Added quantity tracking to ShoppingCartModel & OrderModel
Updated the ShoppingCartModel and OrderModel classes to include a Map of BookModel to Integer, which will be used to track the quantity of each book in the cart. The `bookQuantityMap` field is annotated with `@ElementCollection`, `@CollectionTable`, `@MapKeyJoinColumn`, and `@Column` to define the mapping of the Map in JPA. Also updated the `toString()` method to include the books and their quantities.